# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/user/update": {
      "put": {
        "operationId": "UserController_updateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponseDto" }
              }
            }
          }
        },
        "tags": ["user"],
        "security": [{ "bearer": [] }]
      }
    },
    "/user/find": {
      "get": {
        "operationId": "UserController_findUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponseDto" }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponseDto" }
              }
            }
          }
        },
        "tags": ["user"],
        "security": [{ "bearer": [] }]
      }
    },
    "/user/find-by-email": {
      "get": {
        "operationId": "UserController_findUserByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponseDto" }
              }
            }
          }
        },
        "tags": ["user"],
        "security": [{ "bearer": [] }]
      }
    },
    "/user/list": {
      "get": {
        "operationId": "UserController_listUsers",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of users to take",
            "schema": { "type": "number" }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination",
            "schema": { "type": "string" }
          },
          {
            "name": "sortColumn",
            "required": false,
            "in": "query",
            "description": "Column to sort by",
            "schema": { "type": "string" }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserListDto" }
              }
            }
          }
        },
        "tags": ["user"],
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "info": {
    "title": "Starter API",
    "description": "The Nestjs Starter By @sshahriazz",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "password": { "type": "string" },
          "email_verified": { "type": "boolean" },
          "disable_access": { "type": "boolean" },
          "avatar": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "updated_at": { "format": "date-time", "type": "string" }
        },
        "required": ["id"]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "password": { "type": "string" },
          "email_verified": { "type": "boolean" },
          "disable_access": { "type": "boolean" },
          "avatar": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "updated_at": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "email",
          "first_name",
          "last_name",
          "password",
          "email_verified",
          "disable_access",
          "avatar",
          "created_at",
          "updated_at"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "password": { "type": "string" },
          "email_verified": { "type": "boolean" },
          "disable_access": { "type": "boolean" },
          "avatar": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "updated_at": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "email",
          "first_name",
          "last_name",
          "password",
          "email_verified",
          "disable_access",
          "avatar",
          "created_at",
          "updated_at"
        ]
      },
      "UserListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserDto" }
          },
          "nextCursor": { "type": "string" }
        },
        "required": ["data"]
      },
...
    }
  }
}